requirements.txt
numpy
tensorflow
azure-search-documents
flask